Update behaviour of select-with-search component
Based of some feedback we got on our documents search page we're
making some changes to the select-with-search components behaviour.

This does the following:

1. Removes the "Press to select" text from options when highlighted
2. Updates fuzzy search to match sequential characters at any part of
the string
3. Removes the padding from the right of options where "Press to select"
was located. It's now 5px instead of 100px
4. Apply 'word-break: break-word;' so that results are more readable
5. Increase search result limit to 100

All of these changes bring the component more in line with the previous
behaviour that Select2 exhibits.