Update currently_featured_tab_component to work with featurings
When building this i hadn't realised that there is a features page
for topical events. Instead of features belonging to a feature list,
which in turn belongs to a world locations news or org, featurables
belong to topical events.

They have a lot more in common than different so i'm going to retain
the currently_featured_tab_components core behaviour, but tweak
it so that it works for both features and featurings.

This updates the currently_featured_tab_component to do this. I'm going
to add the two new components which handle rendering the featured
documents for both into their own commits so it's a bit easier to
review.