Ensure Government#edit page handles blank start date gracefully
At the moment, this is blowing up as a model on the validation
expects start date to be present. This updates the validation for
overlapping governments to only run when start_date is present.

It also fixes a bunch of issues with the code itself. Duplicate fields,
values not being set in inputs correctly, label size etc.