Convert reshuffle govspeak to HTML before storing
We don't want to introduce more govspeak rendering knowledge into front-
end apps than necessary.

The reshuffle message used for the latest reshuffle included govspeak.
Convert it to HTML before storing it in the content item.

`bare_govspeak_to_html` was the method I found that introduced the least
amount of HTML, to avoid interfering with the components used in the
rendering app (in this case collections).