Add mapped_specialist_topic_content_id to document collection presenter
This field will be used as part of retiring specialist topics, some of which will be redirected to document collections that have been created to replace them. These document collections will have a mapped_specialist_topic_content_id thatrefers to the specialist topic they are intended to replace.

Having this field present on the document collection means that we will be able to pick out document collections where we can automatically migrate all email subscribers of the specialist topic over to the document collection.

We can remove this field once all specialist topics have been retired.