Present appointed people in the content item
Whitehall currently shows any people who are appointed to certain roles
in a worldwide org in the following way:
- image and details for the person in a primary role, if they exist
- details only for people in secondary or office roles.

In order to render them in government-frontend, we need to have the
list of people and their roles in the content item.

I chose to preserve the secondary and office distinction, in case it
ever becomes relevant when displaying in the rendering app.