Delete attachments from Worldwide Org About pages
In the commit titled "Move attachments from about page to
WorldwideOrganisation" we move attachments from published About
Corporate Information Pages to the corresponding Worldwide Organisation.

In this commit we're updating the tidy-up migration so that as well as
deleting the now redundant About Corporate Information Pages we also
delete any attachments associated with those pages. The `has_many
:attachments` comes from the `Attachable` concern and I've copied enough
of the `Attachment` class to get everything to work.

The output of this destructive migration is very verbose in case we need
to recover any of the data. An example of the output:

Processing DeleteWorldwideOrganisationAboutPages::WorldwideOrganisation 503 - uk-science-innovation-network-in-colombia
  Deleting CorporateInformationPage 923435 - About us
    Deleting FileAttachment 3234866 - UK Science and Innovation Network Country Snapshot - Colombia
  Deleting CorporateInformationPage 1265622 - About us
    Deleting FileAttachment 6017156 - UK Science and Innovation Network Country Snapshot - Colombia
  Deleting CorporateInformationPage 1427071 - About us

I ran this against a reasonably recent version of the Whitehall data in
development and it took about 15 minutes. The summary printed at the end
of the migration was:

== Summary

=== Before

WorldwideOrganisations: 435
About Us Corporate Information Pages: 10742
Attachments: 2602

=== After

WorldwideOrganisations: 435
About Us Corporate Information Pages: 0
Attachments: 0