Update the features for Worldwide Office access information
Once rendering has been removed from Whitehall, these features will fail.

However, they are helpful to ensure that the office is associated with the
correct access information. 

This updates them to check the edit page instead of the public page; this is
now the only page that displays access information.