Update promotional_features_controller#new to render design system
This updates the controller to render the new view and the design
system layout when the user has the preview design system permission.

This adds the new view and form partial for promotional fields.
I've chosen not to add all the changes to the
view in 1 commit just so it's easier to review. I'll squash them
before merge.

As the nested promotional_feature_item partials are reused by the
promotional feature item new and edit endpoints, i've built
them in an object agnostic way by passing an id and name into the
promotional feature items partials. These will be used to construct the
form fields.