Ensure heading size is correct when partials are called
There was a subtle issue when the promotional_feature_item
form fields partial was called from the new and edit pages.

As the title, title_url and summary fields are not contained in a
fieldset they should have a large heading rather than medium.

This adds and additional local method which takes a heading size.