Fix ability to edit translation documents
Previously, locale was not added to the publishing API publish command, meaning the default english document is loaded, and attempting to publish an already published document returns an error 409 unable to publish already published document.
With the fix, publish command should be able to load the correct document with the previous draft created in the PUT command.