Test links to access details in the feature specs
We had to revert (#7760) the work to remove rendering of Worldwide Offices due
issue where we were still using the old url helpers after removing a route.

This wasn't caught by any tests as the relevant code is conditional on an
office having access details.

This adds additional steps to ensure links to access details are intact.