Design updates for the new/edit role appointments page
This updates the design in a few ways:

When the appointment is not a current appointment it renders the start
date and end date within a fieldset with medium sized headings.

When it is a current appointment then it only renders a start date with
a large heading.

It also updates the label text and adds 50px margin between the button
group and the last input.

I've done a tiny bit of refactoring too.