Simplify 'render_design_system' method
We've had an issue a few times where we've accidentally rendered a
template built in the GOV.UK Design System with the Bootstrap
layout.

This is due to pages accidentally being moved into the next release.

This refactor couples view rendering with the get_layout method in
each controller.

If get_layout returns design system then the design_system_view
is rendered. If not then the legacy_view is rendered.