Add shared test which checks admin template is rendered
Now that the logic that renders the design system or bootstrap views
is using the get_layout method to as a parameter, we can be confident
that the view that corresponds to the layout will be rendered.

This commit adds a shared test to all admin controllers that still need
to be ported to the GOV.UK Design System.

This checks for a get_layout method.

If it exists, it creates a user with the "Preview next release"
permission, then iterates through each non-inherited controller
action (new create edit update etc.) and checks that get_layout
returns "admin".

Once pages are moved into the next release the test will fail and it can
be removed.