Ensure integration tests test Boostrap and GDS Implementation
This duplicates a couple of integration tests that failed when we
overrode the preview_design_system? method to always return true
for pages going out in the 1.9 release.

It also fixes an issue on the world location translation edit page
where divs were being closed inside the form despite being opened
before it.

This caused some bizarre behaviour where the id of the world location
news was not being passed to the controller in integration tests despite
working on production.