Add World Location names to the details for Worldwide Organisations
In the content item for Worldwide Organisations, we are currently
including a link to the world location(s) that are associated with this
worldwide organisation.

However the linked content item only includes the location name in
English.

We need the location to be included in the language that the page is
currently being rendered in.

Therefore adding an additional item in `details` that will include the
names in the correct language. We will then map these back to the
`links` using the content ID.