Allow option to layout to make the title, error summary and banner full width
Some of our pages are built in full width. When this is the case,
the title, error summary and banners should render in full width so
the pages look consistent.

This adds this functionality to the design_system layout and applies
it in the statistics_announcements index page and editions index
pages.