Rename consultation responses and consultation Uploader
Response model is only associated with Consultation, therefore should be renamed to something more specific such as ConsultationResponse. This is also to prepare us for introducing CallForEvidenceResponse. We want to avoid having polymorphic relationships where we create ConsultationResponse and CallForEvidenceResponse to inherit Response because this will create 2 layers of polymorphism, including the 2 classes that currently inherit from (Consultation)Response - ConsultationOutome and ConsultationPublicFeedback. To keep things simple and easy to navigate and maintain in the future, we have opted to rename Response
ConsultationResponseFormUploader was renamed since it was an uploader for documents. There was nothing in the uploader that signified anything specific to consultation, and also it was for uploading documents given the file extensions. So rather than calling it a form uploader, we can make it a generic document uploader to reuse for both Consultation and Call For Evidence and anything else that might require a document in a similar way.

Co-authored-by: Minno Dang <minno.dang@digital.cabinet-office.gov.uk>
Co-authored-by: Barbara Slawinska <barbara.slawinska@digital.cabinet-office.gov.uk>