Display an error message when attempting to preview invalid govspeak
Presently, the govspeak preview endpoint on the server returns a 403 forbidden response when invalid govspeak or govspeak premitting XSS attacks is provided. This commit handles the 403 response by displaying an error to the user asking them to ensure the provided govspeak is valid. The error message is removed when the user returns to edit the govspeak.