Don't provide links to preview draft until first successful save
There's an issue that a few publishers have encountered where they
try and share a preview link with colleagues before an edition has
been saved after a new edition has been created.

This occurs in the following situation:

1. a published document exists
2. the publisher creates a new draft document
3. when they arrive on the edit page they don't add a change note
or mark the version as minor.
4. they leave the edit page via exiting the browser or by clicking the
cancel link.
5. they have access to preview links and the sharable preview link
functionality

When a draft is successfully saved, it is pushed downstream to
PublishingAPI and then to Draft Content Store where if finds the
latest draft and overrides it.

This means that until this save occurs the Draft Content Store will
return the latest draft. In this case that would be the draft of the
now published edition.

When the user clicks the 'Preview on website (opens in new tab)' link
they'll hit draft-origin, which will make the call to content store for
the latest draft for that content_id. As stated this will load the
latest draft it knows about which is now out of date.

This means you can do things like add attachments and update the
taxonomy of the latest draft on Whitehall, and click the preview link and
none of these changes will be reflected.

Whitehall only passes an edition to Publishing API when it is valid
which it isn't as there's no change note.

Additionally, the sharable preview link will continue to show, but won't
work for the following reason.

- the new edition is created with a new auth_bypass_id
- the sharable preview link is rendered with a token which is
constructed using the new draft editions auth_bypass_id
- the user takes the link
- the old auth_bypass_id lives on the old draft in the Draft Content
Store
- it attempts to descrypt the token using the old auth_bypass_id
- it doesn't work as the token uses the new auth_bypass_id from
the latest draft edition on Whitehall