Link draft attachments to their parent document
Asset Manager has a concept of linking assets to their parent document.
We use this URL in the CSV Previews to retrieve the metadata about the
parent document.

At the moment, Whitehall only sends the parent document URL for live
assets.

This extends that to send the parent document URL for draft assets too.
Doing this will allow us to offer previews of draft CSV files to users.