Replace the Gem for "govuk_app_config" with JSON logger format version
Branch "Log-GOVUK_REQUEST_ID-in-Json-in-Whitehall" in "alphagov/govuk_app_config" is used so that we can deploy to other environments as integration test.

Co-authored-by: Neamah Al-Selwi <neamah.alselwi@digital.cabinet-office.gov.uk>
Co-authored-by: Tuomas Nylund <tuomas.nylund@digital.cabinet-office.gov.uk>

Replace the Gem for "govuk_app_config" with JSON logger format version

Branch "Log-GOVUK_REQUEST_ID-in-Json-in-Whitehall" in "alphagov/govuk_app_config" is used so that we can deploy to other environments as integration test.

Co-authored-by: Neamah Al-Selwi <neamah.alselwi@digital.cabinet-office.gov.uk>
Co-authored-by: Tuomas Nylund <tuomas.nylund@digital.cabinet-office.gov.uk>