Fix broken links filter on documents search page
At the moment, this filter isn't working correctly. Some link reports
have links in a pending state. As the SQL was only checking that all
the links were in the "ok" state, any edition with a "pending" link
was also being returned.

This updates the SQL to explicitly check whether the link report has a
link with a "broken" or "caution" status.