Add feature flag for Call For Evidence
We decided to add the feature flag check within the edition rules rather than the controller because this would be the one place we need to make a change and it would control both access to the controllers as well as what document types can be created from the drop down in the navigation bar.
Some of the user logins in tests at the moment have deep rooted logins shared across all controller tests, and user permissions needed to be appended to all. We might need to consider refactoring this.

Co-authored-by: Ryan Brown <ryan.brown@digital.cabinet-office.gov.uk>
Co-authored-by: Jonathan Young <jonathan.young@digital.cabinet-office.gov.uk>