Update single image upload component to allow custom title and names
The new/edit topical event pages allow a user to upload a logo for
their topical event.

At the moment, the single image upload defaults the name of the alt text
field to name_you_pass_in[alt_text] and the image field to
name_you_pass_in[image] which works most of the time.

This allows customisation of these fields so we can use
topical_event[logo] etc.

Co-authored-by: farahTW farah.mehmood@digital.cabinet-office.gov.uk