Fix bug with filter for topical event featurings
At the moment, this method returns all featurable editions for a
topical event featurings.

This isn't actually the use case needed. What this needs to do is
take a subset of editions and filter out any that belong to a
topical event featuring.

This allows it to work with the paginator which only renders
15 per page.