Ensure cancel link works on unsuccessful save for new featurings page
When the page saves unsuccessfully the cancel link
doesn't construct the link correctly. When you click it no HTTP request
is made.

This only occurs when an anchor is present. As the PATCH update resource
has the same url the GET index endpoint instead of reloading the page,
it simply enacts the anchors behaviour when you click on it.

As far as i can tell there's no way to force a reload without using JS
or adding a query string. This isn't an ideal situation but will
force a page reload when clicked.