Remove extra about page for British Indian Ocean Territory
We have recently updated the base_path for the "about us" worldwide corporate
information pages. 

Because of this, we have seen a bunch of `UnpublishableInstanceError`s for this
particular worldwide organisation as it has two "about us" pages and
therefore two of the same base_paths. The "about us" page that is not in use
has very minor changes (compared to the one that is) and we have decided as a
team that it can be removed.

We plan to do further work to identify where other worldwide organisations have
multiples of the same type of CIP, including adding validation where it is
missing so that this issue can be prevented in future.