Store asset_manager_id to track Asset Manager assets
Created assets table to save asset_manager_id as an alternative way of
tracking "remote" assets from Asset Manager. Current implementation
relies on the use of legacy_url_path which we are trying to remove.

Knowing the variant of an asset is required in order to differentiate
between the original file and its various carrierwave versions
(e.g. pdf and thumbnail).

We are saving the asset_manager_id outside the AssetManagerAttachmentSetUploadedToWorker
with the assumption that AttachmentData should be available and there is no
need to wait 0.5s.

Co-authored-by: Syed Ali syed.ali@digital.cabinet-office.gov.uk