Block out SVG for Default Organisation News Image.
Currently edit page for Organisation 500 errors when SVG is uploaded since it is trying to render a specific sized version of the image that doesn't exist for an SVG image.
We are blocking out SVG upload since SVG should be only used for charts and graphs and not for these types of default lead images.

Co-Authored-By: Jonathan Young <jonathan.young@digital.cabinet-office.gov.uk>