Create table for review_reminders
This creates a new table which will be used to store details for
sending reminder emails to users to remind them to review their content.

Although it's explicitly coupled to documents at the moment, we could
rework this in the future to be polymorphic so that different content
types can have reminders (topical events etc.).