Build blank document and review_reminder in before action
As we're going to have fields for review_reminder in the new and
edit edition pages, we need to ensure that we build a blank document
and review reminder so that nested attributes will be accepted.

Doing this has caused a few complications:

1. the methods that determined national inapplicability were expecting
document to be nil on a new object so i've updated it to also accept
a new record.
2. the before_validation action that sets up the document wasn't
working correctly. It checked for the presence of a document and if one
wasn't found it assigned the content_id and slug.

I've had to update this to set these values when a document exists
but has no values for them as we have to pass a blank document through
for nested attrs to work.