Fix test after gem update
After bumping the govuk_publishing_components gem we were getting this failure:
`
Error:
Admin::DocumentCollectionGroupsControllerTest#test_GET_#index_does_not_show_a_warning_if_<=_50_memberships_in_collection:
ArgumentError: Last argument was a Hash, which would be used for the assertion message. You probably want this to be a String, or you have the wrong type of arguments.
    test/functional/admin/document_collection_groups_controller_test.rb:49:in `block in <class:DocumentCollectionGroupsControllerTest>'
`
To resolve this we have changed the last argument of the assertion to be a string.