Log uploader model class and ID
We want to test the assumption that the AttachmentData may not always
be available at the point in time where the CreateWhitehallAssetWorker
is invoked. We are thus logging the uploader model class and ID - in
order to track both which models are having their attachments uploaded
and whether or not the ID is present.

The AssetManagerAttachmentSetUploadedToWorker is being called from inside
the AssetManagerCreateWhitehallAssetWorker, 0.5s after making the create
call to Asset Manager, presumably because AttachmentData may not exist.

Nonetheless, we have not seen such an occurrence and would like to simplify
these workers altogether. We also want to extend the behaviour of saving an
Asset to other data types (e.g. ImageData). It is thus preferable if the
logic is outside the UploadedToWorker.