Remove autopublishing of retired service and info pages
Collections is responsible for rendering organisation pages. Some orgs
are allowed to have a link to services and information pages on their
organisation page. See [1]

We are currently retiring service and information pages, and have removed
the link from the following pages:

- Highways England https://github.com/alphagov/collections/pull/3166
- MHRA https://github.com/alphagov/collections/pull/3213
- High Speed Two Limited https://github.com/alphagov/collections/pull/3188
- DVSA https://github.com/alphagov/collections/pull/3188
- Planning Inspectorate https://github.com/alphagov/collections/pull/3165
- Environment Agency https://github.com/alphagov/collections/pull/3170

We now need to remove them from the list of orgs held in whitehall. So that
they are no long republished everytime the org page is updated. See [2]

Future work can remove this legacy code once all service and information
pages have been retired. But until then, we need to manually keep the lists in sync.

1 = https://github.com/alphagov/collections/blob/main/app/presenters/organisations/header_presenter.rb#L113-L120
2 = https://github.com/alphagov/whitehall/blob/122a847b5422a795486bc77a6d96417f19955575/app/controllers/admin/organisations_controller.rb#L80