Ensure statistics_announcements dates render in the correct ordering
There's an issue at the moment with the the ordering of statistics
announcements on the index page. This is occurring due to the release
date ordering not being correctly scoped to the current release date.

I've simplified this query so that it's clear what's going on, but the
important thing is that it joins to the current_release_date scope and
includes it in the initial query so that it's pre-loaded from the db
for when future db calls are made it only returns that for the
statistics_announcement_dates association.