Add support in User model for email override editor permission
This permission will allow users to edit where the email subscription for document collections goes. In the majority of cases, we want email subscribers to be subscribed to updates for the document collection itself. In a small number of cases, HMRC would like the email subscription to be that of a taxonomy topic. This will be the case for a small number of document collections created to replace a specialist topic, and supports the overall retiring of specialist topics.

We will only give this permission to a small number of HMRC editors.