Add test for broken link reports
At the moment, when a publisher filters the documents list for only
those with broken links, an edition is included if a broken link has
ever been reported.

This is incorrect as we should only be filtering on the most recent
broken links report (as a link may have been fixed between runs).

Therefore adding a test that confirms this situation. The bug will be
fixed in a later commit.