Only filter broken links on most recent report
At the moment, when a publisher filters the documents list for only
those with broken links, an edition is included if a broken link has
ever been reported.

This is incorrect as we should only be filtering on the most recent
broken links report (as a link may have been fixed between runs).

This adds a condition to the query to ensure that we are only selecting
the most recent report where an edition has been checked multiple times.