Remove public route for detailed guidance
This route was used solely to get the Rails route helper methods and
should have been removed when doing work to remove `UrlMaker`.

However it was missed as the route was located in a different part of
the routes file to everything else.

Therefore removing it now.

This also allows us to remove `rack_404`.