Ensure reminder_sent_at is set to nil when review_at is updated
When a publisher is emailed their review reminder and update the
review date to a date in the future, we need to ensure we set
reminder_sent_at to nil.

If we don't they won't be sent another email on that date as the rake
task only sends emails out when reminder_sent_at is set to nil.