Fix bug with world news story organisation values not being cleared
You can replicate this bug with the following:

1. Make a new story, fill out the required fields and save
2. Change the news story to be world news story and save without
updating any other fields
3. You see an error that you need to add a worldwide organisation and
world location
4. Update the worldwide organisation and world location
5. You see an error that you can't tag a world news story to an
organisation, please remove organisation (but you have no way of doing
this task as the field is not displayed)

This only occurs on an unsuccessful save occurs. This is because lead
and supporting organisations are grabbed from the database each time the
page is reloaded. They don't live on the edition itself.

We can get round this by manually clearing the organisation fields
select values on page load when the NewsArtcile is a World News Story.