Use asset_manager_id when updating an attachment from a rake task
We are introducing an alternative flow for when AttachmentData has assets.
This is part of a bigger change to remove legacy_url_path and use asset IDs
to communicate with Asset Manager instead. If assets are not present, we default
to using legacy_url_path.

The AssetUpdater that the worker calls has already been changed to use
either legacy path or asset manager IDs.

The code using legacy path will be removed once a migration to populate all
Assets is run and we enable creating Assets for each Attachment for all users.

Co-authored-by: Neamah al Selwi neamah.alselwi@digital.cabinet-office.gov.uk