Attempt to fix flaky test
Add another clear of the mailer queue in setup as it doesn't seem enough to just clear it in teardown.
Existing consultation tests do it in both setup and teardown.
Some questions outstanding why just teardown doesn't work. It implies there's a shared global queue for the concurrent tests, but if so the clearing of queue should also affect test flakiness but it doesn't seem to. Continue keeping an eye.