'Pluck' organisation IDs when that's all we need
The CheckAllOrganisationsLinksWorker was pulling all organisations into
memory and looping through them, when all we really needed was the ID of
each organisation.

To optimise resource usage, I've switched it to use `pluck(:id)`
instead.

There are currently 1182 organisations in Whitehall, so this seems like
a reasonable optimisation to make and should result in lower memory
usage for this job.