Ensure all HTML attachments are withdrawn when parent edition is withdrawn
Currently, the publishing API attachments service listener iterates through the list of an edition's HTML attachments and makes a call to the publishing API to withdraw each attachment, but when making the call it uses the edition's locale. This means that attachments with a different locale value from the edition are not marked as withdrawn. This commit resolves this issue by using the attachment's locale in the publishing API call if it is available, falling back to the default system locale if the attachment locale is not specified.