Test for whodunnit/submitter race condition
We've been having a couple of incidents where users have had their
submitted_by being overriden by users in completely different
departments.

This test attempts to recreate the scenario so that we can
implement a fix. It uses `sleep` to recreate conflict
between threads.

I'm undecided that this is a valuable test to keep long-term. I think
this behaviour is unlikely to regress, so it probably isn't needed once
we've proven the fix works.