Add OrganisationForm JS module
This module dynamically hides and shows the logo and
Non-Departmental Public Body Information fields based on the values
passed into the logo crest and organisation type fields.

It also handles cleaning up the params passed to the controller. That
way if someone moves from custom logo to another selection the
logo field is nulled.

And if someone moves from a non_departmental_public_body to another
type, then the 4 non_departmental_public_body params are nulled.