Allow AssetManagerCreateWhitehallAssetWorker to accept temporary asset related params
These changes are fix for production issue:
https://govuk.sentry.io/issues/4336826185/?project=202259&query=&referrer=issue-stream&statsPeriod=14d&stream_index=3
this issue was introduced when the logic for saving assets was moved to AssetManagerCreateAssetWorker
and the number of argument that was passed to the perform method inside AssetManagerCreateWhitehallAssetWorker
was reduced from 8 to 6.
The changes introduced are temporary to be able to execute one job that is currently failing and retrying.

Co-authored-by: Iris Lau  Iris.lau@digital.cabinet-office.gov.uk