Ignore case when sorting world locations
The consumers of the world location content item rely on the data being
correctly sorted (#8002).

However, case is currently not being ignored which means that lowercase
characters come after all uppercase characters. For example, "UK and the
Commonwealth" is incorrectly sorted to be after "UK Mission to ASEAN".

This updates the sort to ignore case.