Make the Audit Trail thread-safe
This commit fixes a thread-safety bug which has resulted in user actions
being misattributed in document history timelines.

The class attribute `AuditTrail.whodunnit` was used to hold the current
user whose request is being served by the app. This class attribute
value is shared with all Threads on the Process which the app runs in.

This never used to be an issue because the web server used to serve
requests via multiple single-threaded processes. So there was no risk of
the variable being shared between multiple requests. However this recently
changed so Whitehall is now served in a multi-threaded environment. This
has resulted in multiple Threads having shared access to the same
`AuditTrail.whodunnit` variable.

When multiple HTTP requests were served in parallel by the same Process
(in separate Threads), the `AuditTrail.whodunnit` variable would
sometimes be written by one Thread and then overwritten by the second
Thread before Audit Trail records were written. The result was that the
wrong user would sometimes be attributed to a document change.

In this commit, I've introduced a new `Current` model which has a
thread-safe `user` attribute. This has been achieved with the help of
the Rails `ActiveSupport::CurrentAttributes` super class, which ensures
that attribute values are thread-isolated and reset automatically before
and after each request.

So now the current user will be available at `Current.user` rather than
`AuditTrail.whodunnit` throughout the app.