Decouple tests from AuditTrail.whodunnit
Some tests were coupled to the old `AuditTrail.whodunnit` behaviour and
needed adapting to be compatible with the new thread-safe `Current.user`
model.

The most common need was for tests to control which user was considered
to be 'logged in' while certain actions were performed (for example,
when publishing an edition), for the purposes of testing the Audit
Trail.

Where appropriate, I've replaced the manual setting of
`AuditTrail.whodunnit` with use of the `AuditTrail.acting_as` method. I
consider this to be the preferred way to set the current user without
needing to know the details of inner working such as where the current
user is held.

In other cases, I've simply removed the use of `AuditTrail.whodunnit`
altogether because it didn't seem to be needed. Tests still pass without
`whodunnit` being stubbed, for example.

This commit removes all uses of `AuditTrail.whodunnit`, so I've also
removed that attribute from the `AuditTrail` class. It's now gone
entirely.